requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
python-dotenv
psycopg2
